+++ include/coal/collision_data.h
@@ -51,6 +51,7 @@
 #include "coal/timings.h"
 #include "coal/narrowphase/narrowphase_defaults.h"
 #include "coal/logging.h"
+#include "coal/collision_utility.h"
 
 namespace coal {
 
@@ -162,6 +163,34 @@ struct COAL_DLLAPI Contact {
   bool operator!=(const Contact& other) const { return !(*this == other); }
 
   Scalar getDistanceToCollision(const CollisionRequest& request) const;
+
+  /// \brief Prints the contact to the provided output stream.
+  void disp(std::ostream& os, const std::string& prefix = "") const {
+    using namespace std;
+    os << prefix
+       << "  o1 type: " << (o1 ? get_node_type_name(o1->getNodeType()) : "null")
+       << ",\n";
+    os << prefix
+       << "  o2 type: " << (o2 ? get_node_type_name(o2->getNodeType()) : "null")
+       << ",\n";
+    os << prefix << "  b1: " << b1 << ",\n";
+    os << prefix << "  b2: " << b2 << ",\n";
+    os << prefix << "  normal: " << normal.transpose() << ",\n";
+    os << prefix << "  nearest_points[0]: " << nearest_points[0].transpose()
+       << ",\n";
+    os << prefix << "  nearest_points[1]: " << nearest_points[1].transpose()
+       << ",\n";
+    os << prefix << "  pos: " << pos.transpose() << ",\n";
+    os << prefix << "  penetration_depth: " << penetration_depth << ",\n";
+  }
+
+  /// \copydoc disp
+  friend std::ostream& operator<<(std::ostream& os, const Contact& contact) {
+    os << "Contact: {" << "\n";
+    contact.disp(os);
+    os << "}" << "\n";
+    return os;
+  }
 };
 
 struct QueryResult;
@@ -269,6 +298,39 @@ struct COAL_DLLAPI QueryRequest {
            collision_distance_threshold == other.collision_distance_threshold;
     COAL_COMPILER_DIAGNOSTIC_POP
   }
+
+  /// @brief whether two QueryRequest are different or not
+  inline bool operator!=(const QueryRequest& other) const {
+    return !(*this == other);
+  }
+
+  /// @brief Prints the query request to the provided output stream.
+  void disp(std::ostream& os, const std::string& prefix = "") const {
+    using namespace std;
+    os << prefix << "  gjk_initial_guess: "
+       << (gjk_initial_guess == GJKInitialGuess::DefaultGuess
+               ? "DefaultGuess"
+               : (GJKInitialGuess::CachedGuess ? "CachedGuess"
+                                               : "BoundingVolumeGuess"))
+       << ",\n";
+    os << prefix << "  cached_gjk_guess: " << cached_gjk_guess.transpose()
+       << ",\n";
+    os << prefix << "  cached_support_func_guess: "
+       << cached_support_func_guess.transpose() << ",\n";
+    os << prefix << "  gjk_max_iterations: " << gjk_max_iterations << ",\n";
+    os << prefix << "  gjk_tolerance: " << gjk_tolerance << ",\n";
+    os << prefix << "  gjk_variant: " << static_cast<int>(gjk_variant) << ",\n";
+    os << prefix << "  gjk_convergence_criterion: "
+       << static_cast<int>(gjk_convergence_criterion) << ",\n";
+    os << prefix << "  gjk_convergence_criterion_type: "
+       << static_cast<int>(gjk_convergence_criterion_type) << ",\n";
+    os << prefix << "  epa_max_iterations: " << epa_max_iterations << ",\n";
+    os << prefix << "  epa_tolerance: " << epa_tolerance << ",\n";
+    os << prefix << "  enable_timings: " << enable_timings << ",\n";
+    os << prefix
+       << "  collision_distance_threshold: " << collision_distance_threshold
+       << ",\n";
+  }
 };
 
 /// @brief base class for all query results
@@ -285,6 +347,16 @@ struct COAL_DLLAPI QueryResult {
   QueryResult()
       : cached_gjk_guess(Vec3s::Zero()),
         cached_support_func_guess(support_func_guess_t::Constant(-1)) {}
+
+  /// @brief Prints the query result to the provided output stream.
+  void disp(std::ostream& os, const std::string& prefix = "") const {
+    using namespace std;
+    os << prefix << "  cached_gjk_guess: " << cached_gjk_guess.transpose()
+       << ",\n";
+    os << prefix << "  cached_support_func_guess: "
+       << cached_support_func_guess.transpose() << ",\n";
+    os << prefix << "  timings: " << timings.user << ",\n";
+  }
 };
 
 inline void QueryRequest::updateGuess(const QueryResult& result) const {
@@ -379,6 +451,31 @@ struct COAL_DLLAPI CollisionRequest : QueryRequest {
            distance_upper_bound == other.distance_upper_bound;
     COAL_COMPILER_DIAGNOSTIC_POP
   }
+
+  /// @brief whether two CollisionRequest are different or not
+  inline bool operator!=(const CollisionRequest& other) const {
+    return !(*this == other);
+  }
+
+  /// @brief Prints the collision request to the provided output stream.
+  void disp(std::ostream& os, const std::string& prefix = "") const {
+    using namespace std;
+    QueryRequest::disp(os, prefix);
+    os << prefix << "  num_max_contacts: " << num_max_contacts << ",\n";
+    os << prefix << "  enable_contact: " << enable_contact << ",\n";
+    os << prefix << "  security_margin: " << security_margin << ",\n";
+    os << prefix << "  break_distance: " << break_distance << ",\n";
+    os << prefix << "  distance_upper_bound: " << distance_upper_bound << ",\n";
+  }
+
+  /// \copydoc disp
+  friend std::ostream& operator<<(std::ostream& os,
+                                  const CollisionRequest& request) {
+    os << "CollisionRequest: {" << "\n";
+    request.disp(os);
+    os << "}" << "\n";
+    return os;
+  }
 };
 
 inline Scalar Contact::getDistanceToCollision(
@@ -438,6 +535,11 @@ struct COAL_DLLAPI CollisionResult : QueryResult {
            normal == other.normal;
   }
 
+  /// @brief whether two CollisionResult are the same or not
+  bool operator!=(const CollisionResult& other) const {
+    return !(*this == other);
+  }
+
   /// @brief return binary collision result
   bool isCollision() const { return contacts.size() > 0; }
 
@@ -490,6 +592,35 @@ struct COAL_DLLAPI CollisionResult : QueryResult {
   /// @brief reposition Contact objects when fcl inverts them
   /// during their construction.
   void swapObjects();
+
+  /// \brief Prints the result to the provided output stream.
+  void disp(std::ostream& os, const std::string& prefix = "") const {
+    using namespace std;
+    QueryResult::disp(os, prefix);
+    os << prefix << "  distance_lower_bound: " << distance_lower_bound << ",\n";
+    os << prefix << "  normal: " << normal.transpose() << ",\n";
+    os << prefix << "  nearest_points[0]: " << nearest_points[0].transpose()
+       << ",\n";
+    os << prefix << "  nearest_points[1]: " << nearest_points[1].transpose()
+       << ",\n";
+    os << prefix << "  contacts: {" << "\n";
+    for (std::size_t i = 0; i < contacts.size(); ++i) {
+      os << prefix << "    Contact " << i << ": {" << "\n";
+      const Contact& contact = contacts[i];
+      contact.disp(os, prefix + "      ");
+      os << prefix << "    }\n";
+    }
+    os << prefix << "  }\n";
+  }
+
+  /// \copydoc disp
+  friend std::ostream& operator<<(std::ostream& os,
+                                  const CollisionResult& res) {
+    os << "CollisionResult:" << "\n";
+    res.disp(os);
+    os << "}" << "\n";
+    return os;
+  }
 };
 
 /// @brief This structure allows to encode contact patches.
@@ -671,6 +802,9 @@ struct COAL_DLLAPI ContactPatch {
            this->points() == other.points();
   }
 
+  /// @brief Whether two contact patches are different or not.
+  bool operator!=(const ContactPatch& other) const { return !(*this == other); }
+
   /// @brief Whether two contact patches are the same or not.
   /// Checks for different order of the points.
   bool isSame(
@@ -711,6 +845,32 @@ struct COAL_DLLAPI ContactPatch {
     }
 
     return true;
+  }
+
+  /// @brief Prints the contact patch to the provided output stream.
+  void disp(std::ostream& os, const std::string& prefix = "") const {
+    using namespace std;
+    os << prefix << "  tf: {" << "\n";
+    tf.disp(os, prefix + "    ");
+    os << prefix << "  },\n";
+    os << prefix << "  direction: "
+       << (direction == PatchDirection::DEFAULT ? "DEFAULT" : "INVERTED")
+       << ",\n";
+    os << prefix << "  penetration_depth: " << penetration_depth << ",\n";
+    os << prefix << "  points: {" << "\n";
+    for (size_t i = 0; i < this->size(); ++i) {
+      os << prefix << "    Point " << i << ": " << this->getPoint(i).transpose()
+         << ",\n";
+    }
+    os << prefix << "  }\n";
+  }
+
+  /// \copydoc disp
+  friend std::ostream& operator<<(std::ostream& os, const ContactPatch& patch) {
+    os << "ContactPatch: {" << "\n";
+    patch.disp(os);
+    os << "}" << "\n";
+    return os;
   }
 };
 
@@ -834,6 +994,30 @@ struct COAL_DLLAPI ContactPatchRequest {
                other.getNumSamplesCurvedShapes() &&
            this->getPatchTolerance() == other.getPatchTolerance();
   }
+
+  /// @brief Whether two ContactPatchRequest are different or not.
+  bool operator!=(const ContactPatchRequest& other) const {
+    return !(*this == other);
+  }
+
+  /// @brief Prints the contact patch request to the provided output stream.
+  void disp(std::ostream& os, const std::string& prefix = "") const {
+    using namespace std;
+    os << prefix << "  max_num_patch: " << max_num_patch << ",\n";
+    os << prefix
+       << "  num_samples_curved_shapes: " << getNumSamplesCurvedShapes()
+       << ",\n";
+    os << prefix << "  patch_tolerance: " << getPatchTolerance() << ",\n";
+  }
+
+  /// \copydoc disp
+  friend std::ostream& operator<<(std::ostream& os,
+                                  const ContactPatchRequest& request) {
+    os << "ContactPatchRequest: {" << "\n";
+    request.disp(os);
+    os << "}" << "\n";
+    return os;
+  }
 };
 
 /// @brief Result for a contact patch computation.
@@ -873,20 +1057,64 @@ struct COAL_DLLAPI ContactPatchResult {
     this->set(request);
   };
 
+  /// @brief Copy constructor.
+  ContactPatchResult(const ContactPatchResult& other) { *this = other; }
+
+  /// @brief Move constructor.
+  ContactPatchResult(ContactPatchResult&& other) noexcept {
+    *this = std::move(other);
+  }
+
+  /// @brief Copy assignment operator.
+  ContactPatchResult& operator=(const ContactPatchResult& other) {
+    if (this != &other) {
+      this->m_contact_patches_data = other.m_contact_patches_data;
+      this->m_id_available_patch = other.m_id_available_patch;
+      this->m_contact_patches.clear();
+      for (size_t i = 0; i < this->m_id_available_patch; ++i) {
+        this->m_contact_patches.emplace_back(this->m_contact_patches_data[i]);
+      }
+    }
+    return *this;
+  }
+
+  /// @brief Move assignment operator.
+  ContactPatchResult& operator=(ContactPatchResult&& other) noexcept {
+    if (this != &other) {
+      this->m_contact_patches_data = std::move(other.m_contact_patches_data);
+      this->m_id_available_patch = other.m_id_available_patch;
+      this->m_contact_patches.clear();
+      for (size_t i = 0; i < this->m_id_available_patch; ++i) {
+        this->m_contact_patches.emplace_back(this->m_contact_patches_data[i]);
+      }
+    }
+    return *this;
+  }
+
   /// @brief Number of contact patches in the result.
   size_t numContactPatches() const { return this->m_contact_patches.size(); }
 
   /// @brief Returns a new unused contact patch from the internal data vector.
-  ContactPatchRef getUnusedContactPatch() {
+  ContactPatch& getUnusedContactPatch() {
     if (this->m_id_available_patch >= this->m_contact_patches_data.size()) {
       COAL_LOG_WARNING(
           "Trying to get an unused contact patch but all contact patches are "
           "used. Increasing size of contact patches vector, at the cost of a "
           "copy. You should increase `max_num_patch` in the "
-          "`ContactPatchRequest`.");
-      this->m_contact_patches_data.emplace_back(
-          this->m_contact_patches_data.back());
-      this->m_contact_patches_data.back().clear();
+          "`ContactPatchRequest` when setting up `ContactPatchResult`.");
+      bool need_to_reset_references = (this->m_contact_patches_data.size() ==
+                                       this->m_contact_patches_data.capacity());
+
+      this->m_contact_patches_data.emplace_back(ContactPatch());
+
+      if (need_to_reset_references) {
+        // If we have reached the capacity of the data vector, all references
+        // in `m_contact_patches` are invalidated. We need to reset them.
+        this->m_contact_patches.clear();
+        for (size_t i = 0; i < this->m_id_available_patch; ++i) {
+          this->m_contact_patches.emplace_back(this->m_contact_patches_data[i]);
+        }
+      }
     }
     ContactPatch& contact_patch =
         this->m_contact_patches_data[this->m_id_available_patch];
@@ -979,6 +1207,11 @@ struct COAL_DLLAPI ContactPatchResult {
     return true;
   }
 
+  /// @brief Whether two ContactPatchResult are different or not.
+  bool operator!=(const ContactPatchResult& other) const {
+    return !(*this == other);
+  }
+
   /// @brief Repositions the ContactPatch when they get inverted during their
   /// construction.
   void swapObjects() {
@@ -994,6 +1227,30 @@ struct COAL_DLLAPI ContactPatchResult {
         patch.point(i)(0) *= Scalar(-1);  // only invert the x-axis
       }
     }
+  }
+
+  /// @brief Prints the contact patch result to the provided output stream.
+  void disp(std::ostream& os, const std::string& prefix = "") const {
+    using namespace std;
+    os << prefix << "  numContactPatches: " << this->numContactPatches()
+       << ",\n";
+    os << prefix << "  contact_patches: {" << "\n";
+    for (size_t i = 0; i < this->numContactPatches(); ++i) {
+      os << prefix << "  ContactPatch " << i << ": {" << "\n";
+      const ContactPatch& patch = this->getContactPatch(i);
+      patch.disp(os, prefix + "    ");
+      os << prefix << "  }\n";
+    }
+    os << prefix << "  }\n";
+  }
+
+  /// \copydoc disp
+  friend std::ostream& operator<<(std::ostream& os,
+                                  const ContactPatchResult& result) {
+    os << "ContactPatchResult: {" << "\n";
+    result.disp(os);
+    os << "}" << "\n";
+    return os;
   }
 };
 
@@ -1063,6 +1320,30 @@ struct COAL_DLLAPI DistanceRequest : QueryRequest {
            enable_signed_distance == other.enable_signed_distance &&
            rel_err == other.rel_err && abs_err == other.abs_err;
     COAL_COMPILER_DIAGNOSTIC_POP
+  }
+
+  /// @brief whether two DistanceRequest are different or not
+  inline bool operator!=(const DistanceRequest& other) const {
+    return !(*this == other);
+  }
+
+  /// @brief Prints the distance request to the provided output stream.
+  void disp(std::ostream& os, const std::string& prefix = "") const {
+    using namespace std;
+    QueryRequest::disp(os, prefix);
+    os << prefix << "  enable_signed_distance: " << enable_signed_distance
+       << ",\n";
+    os << prefix << "  rel_err: " << rel_err << ",\n";
+    os << prefix << "  abs_err: " << abs_err << ",\n";
+  }
+
+  /// \copydoc disp
+  friend std::ostream& operator<<(std::ostream& os,
+                                  const DistanceRequest& request) {
+    os << "DistanceRequest: {" << "\n";
+    request.disp(os);
+    os << "}" << "\n";
+    return os;
   }
 };
 
@@ -1184,6 +1465,40 @@ struct COAL_DLLAPI DistanceResult : QueryResult {
     //    *other.o2;
 
     return is_same;
+  }
+
+  /// @brief whether two DistanceResult are different or not
+  inline bool operator!=(const DistanceResult& other) const {
+    return !(*this == other);
+  }
+
+  /// @brief Prints the distance result to the provided output stream.
+  void disp(std::ostream& os, const std::string& prefix = "") const {
+    using namespace std;
+    QueryResult::disp(os, prefix);
+    os << prefix
+       << "  o1: " << (o1 ? get_node_type_name(o1->getNodeType()) : "null")
+       << ",\n";
+    os << prefix
+       << "  o2: " << (o2 ? get_node_type_name(o2->getNodeType()) : "null")
+       << ",\n";
+    os << prefix << "  b1: " << b1 << ",\n";
+    os << prefix << "  b2: " << b2 << ",\n";
+    os << prefix << "  min_distance: " << min_distance << ",\n";
+    os << prefix << "  normal: " << normal.transpose() << ",\n";
+    os << prefix << "  nearest_points[0]: " << nearest_points[0].transpose()
+       << ",\n";
+    os << prefix << "  nearest_points[1]: " << nearest_points[1].transpose()
+       << ",\n";
+  }
+
+  /// \copydoc disp
+  friend std::ostream& operator<<(std::ostream& os,
+                                  const DistanceResult& result) {
+    os << "DistanceResult: {" << "\n";
+    result.disp(os);
+    os << "}" << "\n";
+    return os;
   }
 };